Improve(CI/CD) Test qp: load data only once, adapt after changes in data_loader. Add requirements.txt, change GitHubActions yml file
