+++ .github/workflows/run-tests.yml
@@ -30,4 +30,4 @@ jobs:
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Test with unittest
       run: |
-        python -m unittest test/test_*
+        python test/test_quadratic_program.py
