+++ requirements.txt
@@ -1,0 +1,7 @@
+cvxopt==1.3.2
+matplotlib==3.9.2
+numpy==2.1.3
+pandas==2.2.3
+qpsolvers==4.4.0
+scipy==1.14.1
+statsmodels==0.14.4
